fixed StateMachine test, updated StateMachine, added more StateMachine tests, fixed LinkedList comparison, added a LinkedList comparison test, updated '*[name].get()' to '[name].ref()'

+++ engine/Engine/src/engine/application/GameWindow.cpp
@@ -63,7 +63,7 @@ namespace Ghurund::Engine {
     }
 
     void GameWindow::update() {
-        timeParameter->Value = Timer.ScaledTime;
+        timeParameter->Value = (float)Timer.ScaledTime;
         layers.update(Timer.TimeMs);
         Input->dispatchGamepadEvents(Timer.TimeMs, *this);
         actionMapping.executeDispatches();
